Put ipc renderer and associate handlers to homeWin
